feat(refactor): split language risk pattern emitters into separate modules for better maintainability
