+++ src/audits/code_quality/language_risk/go.rs
@@ -1,5 +1,8 @@
-use crate::findings::types::Severity;
+use super::{line_of, node_text, push_pattern_finding, snippet_of};
+use crate::findings::types::{Finding, Severity};
+use crate::scan::facts::FileFacts;
 use std::path::Path;
+use tree_sitter::Node;
 
 // ── Go ────────────────────────────────────────────────────────────────────────
 
@@ -65,5 +68,53 @@ impl GoRiskPattern {
 
     pub(super) fn base_severity(self) -> Severity {
         Severity::Medium
+    }
+}
+
+/// Emits Go runtime-risk findings from the syntax tree: `panic(...)`,
+/// `log.Fatal`/`log.Fatalf`, and `os.Exit`.
+pub(super) fn emit_go_node(
+    node: Node<'_>,
+    content: &str,
+    path: &Path,
+    file: &FileFacts,
+    findings: &mut Vec<Finding>,
+) {
+    if node.kind() != "call_expression" {
+        return;
+    }
+    let Some(function) = node.child_by_field_name("function") else {
+        return;
+    };
+
+    let pattern = match function.kind() {
+        // `panic(...)`
+        "identifier" if node_text(function, content) == Some("panic") => Some(GoRiskPattern::Panic),
+        // `pkg.Fn(...)` — `log.Fatal`/`log.Fatalf` and `os.Exit`.
+        "selector_expression" => {
+            let package = function
+                .child_by_field_name("operand")
+                .and_then(|n| node_text(n, content));
+            let method = function
+                .child_by_field_name("field")
+                .and_then(|n| node_text(n, content));
+            match (package, method) {
+                (Some("log"), Some("Fatal" | "Fatalf")) => Some(GoRiskPattern::LogFatal),
+                (Some("os"), Some("Exit")) => Some(GoRiskPattern::OsExit),
+                _ => None,
+            }
+        }
+        _ => None,
+    };
+
+    if let Some(pattern) = pattern {
+        push_pattern_finding(
+            &pattern,
+            path,
+            line_of(node),
+            &snippet_of(node, content),
+            file,
+            findings,
+        );
     }
 }
